Add config service to testts
