+++ server/test/src/main/java/io/deephaven/server/test/TestAuthModule.java
@@ -5,15 +5,20 @@ package io.deephaven.server.test;
 
 import dagger.Module;
 import dagger.Provides;
+import io.deephaven.auth.AnonymousAuthenticationHandler;
 import io.deephaven.auth.AuthenticationRequestHandler;
 import io.deephaven.auth.BasicAuthMarshaller;
 import io.deephaven.auth.AuthContext;
+import org.apache.arrow.flight.auth2.Auth2Constants;
 
 import javax.inject.Singleton;
+import java.nio.ByteBuffer;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.UUID;
 
 @Module
 public class TestAuthModule {
@@ -45,9 +50,41 @@ public class TestAuthModule {
     @Provides
     @Singleton
     public Map<String, AuthenticationRequestHandler> bindAuthHandlerMap(BasicAuthMarshaller basicAuthMarshaller) {
-        // note this is mutable
-        HashMap<String, AuthenticationRequestHandler> map = new HashMap<>();
+        Map<String, AuthenticationRequestHandler> map = new HashMap<>();
         map.put(basicAuthMarshaller.getAuthType(), basicAuthMarshaller);
-        return map;
+        AnonymousAuthenticationHandler anonymous = new AnonymousAuthenticationHandler();
+        map.put(anonymous.getAuthType(), anonymous);
+        map.put(FakeBearer.INSTANCE.getAuthType(), FakeBearer.INSTANCE);
+        return Collections.unmodifiableMap(map);
+    }
+
+    public enum FakeBearer implements AuthenticationRequestHandler {
+        INSTANCE;
+
+        public static final String TOKEN = UUID.randomUUID().toString();
+
+        @Override
+        public String getAuthType() {
+            return Auth2Constants.BEARER_PREFIX.trim();
+        }
+
+        @Override
+        public Optional<AuthContext> login(long protocolVersion, ByteBuffer payload,
+                HandshakeResponseListener listener) {
+            return Optional.empty();
+        }
+
+        @Override
+        public Optional<AuthContext> login(String payload, MetadataResponseListener listener) {
+            if (payload.equals(TOKEN)) {
+                return Optional.of(new AuthContext.SuperUser());
+            }
+            return Optional.empty();
+        }
+
+        @Override
+        public void initialize(String targetUrl) {
+            // do nothing
+        }
     }
 }
